requirements.txt
requests
psycopg2-binary
apscheduler
python-dotenv
